CG-0MP00BL1K008M124: Decompose LostCitiesScene into helper classes
Extract responsibilities from 1944-line LostCitiesScene.ts into composable helpers:
- LostCitiesConstants.ts: shared layout, styling, timing, and audio constants
- LostCitiesRenderer.ts: UI creation and refresh logic
- LostCitiesAnimator.ts: card movement and AI animation helpers
- LostCitiesOverlayManager.ts: round and match summary overlays
- LostCitiesTooltipManager.ts: expedition scoring tooltip display
- LostCitiesReplayController.ts: replay mode state injection
- LostCitiesTurnController.ts: turn flow, input handling, and AI execution

LostCitiesScene.ts reduced from 1944 lines to 294 lines.
All tests pass (npm test and npm run build verified).
